+++ CHANGELOG.md
@@ -39,6 +39,12 @@ All notable changes to this project will be documented in this file, in reverse 
 - [#113](https://github.com/zendframework/zend-diactoros/pull/113) fixes an
   issue in the response serializer, ensuring that the status code in the
   deserialized response is an integer.
+- [#115](https://github.com/zendframework/zend-diactoros/pull/115) fixes an
+  issue in the various text-basd response types (`TextResponse`, `HtmlResponse`,
+  and `JsonResponse`); due to the fact that the constructor was not
+  rewinding the message body stream, `getContents()` was thus returning `null`,
+  as the pointer was at the end of the stream. The constructor now rewinds the
+  stream after populating it in the constructor.
 
 ## 1.3.0 - 2015-12-15